fix(apps): define in apps.py for proper registration
Django's convention is to look in the apps.py files for app configs

the OAuthAppConfig.ready() function was not being called from __init__.py
