--- benefits/eligibility/__init__.py
@@ -1,10 +1,0 @@
-"""
-The eligibility application: Verifies eligibility for benefits.
-"""
-from django.apps import AppConfig
-
-
-class EligibilityAppConfig(AppConfig):
-    name = "benefits.eligibility"
-    label = "eligibility"
-    verbose_name = "Eligibility Verification"
