+++ benefits/enrollment/apps.py
@@ -1,0 +1,10 @@
+"""
+The enrollment application: Allows user to enroll payment device for benefits.
+"""
+from django.apps import AppConfig
+
+
+class EnrollmentAppConfig(AppConfig):
+    name = "benefits.enrollment"
+    label = "enrollment"
+    verbose_name = "Benefits Enrollment"
